// src/enum/router.enum.ts
export enum RouterEnum {
  HOME = 'HOME',
  RIFFLE_DETAIL = 'RIFFLE_DETAIL',

  MAGAZINE_LIST = 'MAGAZINE_LIST',
  MAGAZINE_DETAIL = 'MAGAZINE_DETAIL',

  OPTIC_LIST = 'OPTIC_LIST',
  OPTIC_DETAIL = 'OPTIC_DETAIL',

  OPTIC_COLLAR_LIST = 'OPTIC_COLLAR_LIST',
  OPTIC_COLLAR_DETAIL = 'OPTIC_COLLAR_DETAIL',

  AMMUNITION_LIST = 'AMMUNITION_LIST',
  AMMUNITION_DETAIL = 'AMMUNITION_DETAIL',

  HANDGUN_NEW = 'HANDGUN_NEW',
  RIFFLE_NEW = 'RIFFLE_NEW',
  RIFFLE_EDIT = 'RIFFLE_EDIT',
  RIFFLE_LIST = 'RIFFLE_LIST',

  ADMIN_RIFFLE_DETAIL = 'ADMIN_RIFFLE_DETAIL',
  ADMIN_HANDGUN_DETAIL = 'ADMIN_HANDGUN_DETAIL',
  HANDGUN_EDIT = 'HANDGUN_EDIT',
  HANDGUN_LIST = 'HANDGUN_LIST',

  FACTORY_LIST = 'FACTORY_LIST',
  FACTORY_EDIT = 'FACTORY_EDIT',
  FACTORY_NEW = 'FACTORY_NEW',

  RDS_LIST = 'RDS_LIST',
  RDS_DETAIL = 'RDS_DETAIL',

  ADMIN_DASHBOARD = 'ADMIN_DASHBOARD',

  OPTIC_COLLAR_NEW = 'OPTIC_COLLAR_NEW',
  OPTIC_COLLAR_EDIT = 'OPTIC_COLLAR_EDIT',
  AMMUNITION_NEW = 'AMMUNITION_NEW',
  AMMUNITION_EDIT = 'AMMUNITION_EDIT',
  MAGAZINE_EDIT = 'MAGAZINE_EDIT',
  MAGAZINE_NEW = 'MAGAZINE_NEW',

  OPTIC_EDIT = 'OPTIC_EDIT',
  OPTIC_NEW = 'OPTIC_NEW',

  RDS_EDIT = 'RDS_EDIT',
  RDS_NEW = 'RDS_NEW',

  COLOR_NEW = 'COLOR_NEW',
  MATERIAL_NEW = 'MATERIAL_NEW',
  HEAD_TYPE_NEW = 'HEAD_TYPE_NEW',
  BODY_TYPE_NEW = 'BODY_TYPE_NEW',
  THREADED_SIZE_NEW = 'THREADED_SIZE_NEW',
  WEAPON_TYPE_NEW = 'WEAPON_TYPE_NEW',
  CALIBER_NEW = 'CALIBER_NEW',

  INVOICE_NEW = 'INVOICE_NEW',
  INVOICE_EDIT = 'INVOICE_EDIT',
  INVOICE_LIST = 'INVOICE_LIST',
  INVOICE_DETAIL = 'INVOICE_DETAIL',

  NOTFOUND_404 = 'NOTFOUND_404',
  UNAUTHORIZED = 'UNAUTHORIZED'
}